build: refresh js files
