+++ lib/analyze-action-env.test.js
@@ -42,6 +42,7 @@ const actionsUtil = __importStar(require("./actions-util"));
 const analyze = __importStar(require("./analyze"));
 const api = __importStar(require("./api-client"));
 const configUtils = __importStar(require("./config-utils"));
+const gitUtils = __importStar(require("./git-utils"));
 const statusReport = __importStar(require("./status-report"));
 const testing_utils_1 = require("./testing-utils");
 const util = __importStar(require("./util"));
@@ -61,7 +62,7 @@ const util = __importStar(require("./util"));
             .stub(statusReport, "createStatusReportBase")
             .resolves({});
         sinon.stub(statusReport, "sendStatusReport").resolves();
-        sinon.stub(actionsUtil, "isAnalyzingDefaultBranch").resolves(true);
+        sinon.stub(gitUtils, "isAnalyzingDefaultBranch").resolves(true);
         const gitHubVersion = {
             type: util.GitHubVariant.DOTCOM,
         };
