+++ lib/analyze-action-input.test.js
@@ -42,6 +42,7 @@ const actionsUtil = __importStar(require("./actions-util"));
 const analyze = __importStar(require("./analyze"));
 const api = __importStar(require("./api-client"));
 const configUtils = __importStar(require("./config-utils"));
+const gitUtils = __importStar(require("./git-utils"));
 const statusReport = __importStar(require("./status-report"));
 const testing_utils_1 = require("./testing-utils");
 const util = __importStar(require("./util"));
@@ -77,7 +78,7 @@ const util = __importStar(require("./util"));
         optionalInputStub.withArgs("cleanup-level").returns("none");
         optionalInputStub.withArgs("expect-error").returns("false");
         sinon.stub(api, "getGitHubVersion").resolves(gitHubVersion);
-        sinon.stub(actionsUtil, "isAnalyzingDefaultBranch").resolves(true);
+        sinon.stub(gitUtils, "isAnalyzingDefaultBranch").resolves(true);
         (0, testing_utils_1.setupActionsVars)(tmpDir, tmpDir);
         (0, testing_utils_1.mockFeatureFlagApiEndpoint)(200, {});
         process.env["CODEQL_THREADS"] = "1";
