+++ lib/codeql.js
@@ -55,6 +55,7 @@ const cli_errors_1 = require("./cli-errors");
 const doc_url_1 = require("./doc-url");
 const environment_1 = require("./environment");
 const feature_flags_1 = require("./feature-flags");
+const git_utils_1 = require("./git-utils");
 const languages_1 = require("./languages");
 const setupCodeql = __importStar(require("./setup-codeql"));
 const tools_features_1 = require("./tools-features");
@@ -830,7 +831,7 @@ async function getTrapCachingExtractorConfigArgsForLang(config, language) {
     const cacheDir = config.trapCaches[language];
     if (cacheDir === undefined)
         return [];
-    const write = await (0, actions_util_1.isAnalyzingDefaultBranch)();
+    const write = await (0, git_utils_1.isAnalyzingDefaultBranch)();
     return [
         `-O=${language}.trap.cache.dir=${cacheDir}`,
         `-O=${language}.trap.cache.bound=${TRAP_CACHE_SIZE_MB}`,
