+++ lib/database-upload.js
@@ -38,6 +38,7 @@ const fs = __importStar(require("fs"));
 const actionsUtil = __importStar(require("./actions-util"));
 const api_client_1 = require("./api-client");
 const codeql_1 = require("./codeql");
+const gitUtils = __importStar(require("./git-utils"));
 const util = __importStar(require("./util"));
 const util_1 = require("./util");
 async function uploadDatabases(repositoryNwo, config, apiDetails, logger) {
@@ -55,7 +56,7 @@ async function uploadDatabases(repositoryNwo, config, apiDetails, logger) {
         logger.debug("Not running against github.com or GHEC-DR. Skipping upload.");
         return;
     }
-    if (!(await actionsUtil.isAnalyzingDefaultBranch())) {
+    if (!(await gitUtils.isAnalyzingDefaultBranch())) {
         // We only want to upload a database if we are analyzing the default branch.
         logger.debug("Not analyzing default branch. Skipping upload.");
         return;
@@ -79,7 +80,7 @@ async function uploadDatabases(repositoryNwo, config, apiDetails, logger) {
             const bundledDb = await (0, util_1.bundleDb)(config, language, codeql, language);
             const bundledDbSize = fs.statSync(bundledDb).size;
             const bundledDbReadStream = fs.createReadStream(bundledDb);
-            const commitOid = await actionsUtil.getCommitOid(actionsUtil.getRequiredInput("checkout_path"));
+            const commitOid = await gitUtils.getCommitOid(actionsUtil.getRequiredInput("checkout_path"));
             try {
                 await client.request(`POST /repos/:owner/:repo/code-scanning/codeql/databases/:language?name=:name&commit_oid=:commit_oid`, {
                     baseUrl: uploadsBaseUrl,
