+++ lib/database-upload.test.js
@@ -44,6 +44,7 @@ const actionsUtil = __importStar(require("./actions-util"));
 const apiClient = __importStar(require("./api-client"));
 const codeql_1 = require("./codeql");
 const database_upload_1 = require("./database-upload");
+const gitUtils = __importStar(require("./git-utils"));
 const languages_1 = require("./languages");
 const testing_utils_1 = require("./testing-utils");
 const util_1 = require("./util");
@@ -85,7 +86,7 @@ async function mockHttpRequests(databaseUploadStatusCode) {
             .stub(actionsUtil, "getRequiredInput")
             .withArgs("upload-database")
             .returns("false");
-        sinon.stub(actionsUtil, "isAnalyzingDefaultBranch").resolves(true);
+        sinon.stub(gitUtils, "isAnalyzingDefaultBranch").resolves(true);
         const loggedMessages = [];
         await (0, database_upload_1.uploadDatabases)(testRepoName, getTestConfig(tmpDir), testApiDetails, (0, testing_utils_1.getRecordingLogger)(loggedMessages));
         t.assert(loggedMessages.find((v) => v.type === "debug" &&
@@ -100,7 +101,7 @@ async function mockHttpRequests(databaseUploadStatusCode) {
             .stub(actionsUtil, "getRequiredInput")
             .withArgs("upload-database")
             .returns("true");
-        sinon.stub(actionsUtil, "isAnalyzingDefaultBranch").resolves(true);
+        sinon.stub(gitUtils, "isAnalyzingDefaultBranch").resolves(true);
         const config = getTestConfig(tmpDir);
         config.gitHubVersion = { type: util_1.GitHubVariant.GHES, version: "3.0" };
         const loggedMessages = [];
@@ -117,7 +118,7 @@ async function mockHttpRequests(databaseUploadStatusCode) {
             .stub(actionsUtil, "getRequiredInput")
             .withArgs("upload-database")
             .returns("true");
-        sinon.stub(actionsUtil, "isAnalyzingDefaultBranch").resolves(false);
+        sinon.stub(gitUtils, "isAnalyzingDefaultBranch").resolves(false);
         const loggedMessages = [];
         await (0, database_upload_1.uploadDatabases)(testRepoName, getTestConfig(tmpDir), testApiDetails, (0, testing_utils_1.getRecordingLogger)(loggedMessages));
         t.assert(loggedMessages.find((v) => v.type === "debug" &&
@@ -131,7 +132,7 @@ async function mockHttpRequests(databaseUploadStatusCode) {
             .stub(actionsUtil, "getRequiredInput")
             .withArgs("upload-database")
             .returns("true");
-        sinon.stub(actionsUtil, "isAnalyzingDefaultBranch").resolves(true);
+        sinon.stub(gitUtils, "isAnalyzingDefaultBranch").resolves(true);
         await mockHttpRequests(500);
         (0, codeql_1.setCodeQL)({
             async databaseBundle(_, outputFilePath) {
@@ -152,7 +153,7 @@ async function mockHttpRequests(databaseUploadStatusCode) {
             .stub(actionsUtil, "getRequiredInput")
             .withArgs("upload-database")
             .returns("true");
-        sinon.stub(actionsUtil, "isAnalyzingDefaultBranch").resolves(true);
+        sinon.stub(gitUtils, "isAnalyzingDefaultBranch").resolves(true);
         await mockHttpRequests(201);
         (0, codeql_1.setCodeQL)({
             async databaseBundle(_, outputFilePath) {
@@ -172,7 +173,7 @@ async function mockHttpRequests(databaseUploadStatusCode) {
             .stub(actionsUtil, "getRequiredInput")
             .withArgs("upload-database")
             .returns("true");
-        sinon.stub(actionsUtil, "isAnalyzingDefaultBranch").resolves(true);
+        sinon.stub(gitUtils, "isAnalyzingDefaultBranch").resolves(true);
         const databaseUploadSpy = await mockHttpRequests(201);
         (0, codeql_1.setCodeQL)({
             async databaseBundle(_, outputFilePath) {
