+++ lib/status-report.js
@@ -45,6 +45,7 @@ const actions_util_1 = require("./actions-util");
 const api_client_1 = require("./api-client");
 const doc_url_1 = require("./doc-url");
 const environment_1 = require("./environment");
+const git_utils_1 = require("./git-utils");
 const util_1 = require("./util");
 var ActionName;
 (function (ActionName) {
@@ -125,7 +126,7 @@ function setJobStatusIfUnsuccessful(actionStatus) {
 async function createStatusReportBase(actionName, status, actionStartedAt, config, diskInfo, logger, cause, exception) {
     try {
         const commitOid = (0, actions_util_1.getOptionalInput)("sha") || process.env["GITHUB_SHA"] || "";
-        const ref = await (0, actions_util_1.getRef)();
+        const ref = await (0, git_utils_1.getRef)();
         const jobRunUUID = process.env[environment_1.EnvVar.JOB_RUN_UUID] || "";
         const workflowRunID = (0, actions_util_1.getWorkflowRunID)();
         const workflowRunAttempt = (0, actions_util_1.getWorkflowRunAttempt)();
