+++ lib/trap-caching.test.js
@@ -45,6 +45,7 @@ const actionsUtil = __importStar(require("./actions-util"));
 const apiClient = __importStar(require("./api-client"));
 const codeql_1 = require("./codeql");
 const feature_flags_1 = require("./feature-flags");
+const gitUtils = __importStar(require("./git-utils"));
 const languages_1 = require("./languages");
 const logging_1 = require("./logging");
 const testing_utils_1 = require("./testing-utils");
@@ -111,7 +112,7 @@ function getTestConfigWithTempDir(tempDir) {
 (0, ava_1.default)("check flags for JS, analyzing default branch", async (t) => {
     await util.withTmpDir(async (tmpDir) => {
         const config = getTestConfigWithTempDir(tmpDir);
-        sinon.stub(actionsUtil, "isAnalyzingDefaultBranch").resolves(true);
+        sinon.stub(gitUtils, "isAnalyzingDefaultBranch").resolves(true);
         const result = await (0, codeql_1.getTrapCachingExtractorConfigArgsForLang)(config, languages_1.Language.javascript);
         t.deepEqual(result, [
             `-O=javascript.trap.cache.dir=${path.resolve(tmpDir, "jsCache")}`,
@@ -123,7 +124,7 @@ function getTestConfigWithTempDir(tempDir) {
 (0, ava_1.default)("check flags for all, not analyzing default branch", async (t) => {
     await util.withTmpDir(async (tmpDir) => {
         const config = getTestConfigWithTempDir(tmpDir);
-        sinon.stub(actionsUtil, "isAnalyzingDefaultBranch").resolves(false);
+        sinon.stub(gitUtils, "isAnalyzingDefaultBranch").resolves(false);
         const result = await (0, codeql_1.getTrapCachingExtractorConfigArgs)(config);
         t.deepEqual(result, [
             `-O=javascript.trap.cache.dir=${path.resolve(tmpDir, "jsCache")}`,
@@ -144,7 +145,7 @@ function getTestConfigWithTempDir(tempDir) {
 (0, ava_1.default)("upload cache key contains right fields", async (t) => {
     const loggedMessages = [];
     const logger = (0, testing_utils_1.getRecordingLogger)(loggedMessages);
-    sinon.stub(actionsUtil, "isAnalyzingDefaultBranch").resolves(true);
+    sinon.stub(gitUtils, "isAnalyzingDefaultBranch").resolves(true);
     sinon.stub(util, "tryGetFolderBytes").resolves(999_999_999);
     const stubSave = sinon.stub(cache, "saveCache");
     process.env.GITHUB_SHA = "somesha";
@@ -159,7 +160,7 @@ function getTestConfigWithTempDir(tempDir) {
         const loggedMessages = [];
         const logger = (0, testing_utils_1.getRecordingLogger)(loggedMessages);
         sinon.stub(actionsUtil, "getTemporaryDirectory").returns(tmpDir);
-        sinon.stub(actionsUtil, "isAnalyzingDefaultBranch").resolves(false);
+        sinon.stub(gitUtils, "isAnalyzingDefaultBranch").resolves(false);
         const stubRestore = sinon.stub(cache, "restoreCache").resolves("found");
         const eventFile = path.resolve(tmpDir, "event.json");
         process.env.GITHUB_EVENT_NAME = "pull_request";
@@ -185,8 +186,8 @@ function getTestConfigWithTempDir(tempDir) {
     await util.withTmpDir(async (tmpDir) => {
         // This config specifies that we are analyzing JavaScript and Ruby, but not Swift.
         const config = getTestConfigWithTempDir(tmpDir);
-        sinon.stub(actionsUtil, "getRef").resolves("refs/heads/main");
-        sinon.stub(actionsUtil, "isAnalyzingDefaultBranch").resolves(true);
+        sinon.stub(gitUtils, "getRef").resolves("refs/heads/main");
+        sinon.stub(gitUtils, "isAnalyzingDefaultBranch").resolves(true);
         const listStub = sinon.stub(apiClient, "listActionsCaches").resolves([
             // Should be kept, since it's not relevant to CodeQL. In reality, the API shouldn't return
             // this in the first place, but this is a defensive check.
